Ensure that controllerreferences are checked before object deletion

+++ controllers/utils/generic_object.go
@@ -153,7 +153,8 @@ func ReconcileGeneric[T client.Object](ctx context.Context, c client.Client, own
 	return existingObj, false, nil
 }
 
-// DeleteGeneric deletes a generic object of type T from the cluster
+// DeleteGeneric deletes a generic object of type T from the cluster.
+// If owner is non-nil, the object is only deleted if it has an owner reference pointing to owner.
 /*
 T must be a pointer to a k8s object, e.g., *corev1.ConfigMap
 
@@ -161,7 +162,7 @@ Returns:
 - a boolean flag indicating whether the object exists/existed on the cluster
 - any error
 */
-func DeleteGeneric[T client.Object](ctx context.Context, c client.Client, resourceKind string, config GenericConfig) (bool, error) {
+func DeleteGeneric[T client.Object](ctx context.Context, c client.Client, owner metav1.Object, resourceKind string, config GenericConfig) (bool, error) {
 	// Allocate a new pointer to the struct that T points to, and cast to T
 	// e.g., replaces existingRoute := &routev1.Route{}
 	var existingObj T
@@ -179,13 +180,22 @@ func DeleteGeneric[T client.Object](ctx context.Context, c client.Client, resour
 	} else if err != nil {
 		LogErrorRetrieving(ctx, err, resourceKind, *config.Name, *config.Namespace)
 		return true, err
-	} else {
-		err = c.Delete(ctx, existingObj)
-		if err != nil {
-			LogErrorVerb(ctx, err, "deleting", resourceKind, *config.Name, *config.Namespace)
-			return true, err
-		} else {
-			return true, nil
-		}
 	}
+
+	if owner != nil && !isControlledBy(existingObj, owner) {
+		return false, nil
+	}
+
+	err = c.Delete(ctx, existingObj)
+	if err != nil {
+		LogErrorVerb(ctx, err, "deleting", resourceKind, *config.Name, *config.Namespace)
+		return true, err
+	}
+	return true, nil
+}
+
+// isControlledBy reports whether obj's controller owner reference points to the proviced owner.
+func isControlledBy(obj metav1.Object, owner metav1.Object) bool {
+	ref := metav1.GetControllerOf(obj)
+	return ref != nil && ref.UID == owner.GetUID()
 }
